Provider abstraction and Tavily extraction
