+++ src/lib/tools/trend-search/providers/serper.ts
@@ -1,0 +1,6 @@
+import type { RawSearchResult } from "~/lib/tools/trend-search/types";
+
+/** Placeholder: to be implemented in task 2.1. */
+export async function callSerper(_query: string): Promise<RawSearchResult[]> {
+    return [];
+}
